Move GridContextMenu classes to their own package
This is to revert a backwards incompatible change,
and fix OSGi compatibility for the plaftorm.
