+++ vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/grid/contextmenu/GridLoadsItemsIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.contextmenu;
+package com.vaadin.flow.component.grid.contextmenu;
 
 import java.util.Arrays;
 import java.util.List;
