+++ vaadin-grid-flow/src/test/java/com/vaadin/flow/component/grid/contextmenu/GridSubMenuTest.java
@@ -13,8 +13,9 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.contextmenu;
+package com.vaadin.flow.component.grid.contextmenu;
 
+import com.vaadin.flow.component.contextmenu.MenuManager;
 import org.junit.Test;
 import org.mockito.Mockito;
